% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_interactions.R
\docType{data}
\name{drug_interactions}
\alias{drug_interactions}
\title{Drug interactions}
\format{A data frame with 38116 rows and 6 columns
\describe{
  \item{entrez_gene_symbol}{entrez gene symbol}
  \item{gene_long_name}{gene long name}
  \item{interaction_claim_source}{interaction claim source}
  \item{interaction_types}{interaction types}
  \item{drug_name}{drug name}
  \item{drug_primary_name}{drug primary name}
}}
\source{
The "Interactions TSV" at \url{http://dgidb.genome.wustl.edu/downloads}
}
\usage{
drug_interactions
}
\description{
List of drug names and the genes that they target, along with information about how they interact and source of information
}
\keyword{datasets}

