% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_trials.R
\docType{data}
\name{gene_trials}
\alias{gene_trials}
\title{Drug trial information regarding target genes}
\format{A data frame with 19085 rows and 5 columns
\describe{
	\item{drug_proxy}{Binary value for whether used as drug proxy}
	\item{MeSH1}{MeSH term 1}
	\item{MeSH2}{Top MeSH term}
	\item{drug_phase}{Phase of drug development}
	\item{approved}{Approved in EU or US}
 }}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/26121088}
}
\usage{
gene_trials
}
\description{
The drug trial information comes from PMID 26121088 supplementary table 13.
For a number of genes information is provided about diseases for which they are targeted, and the status of the trials in which they were tested.
}
\keyword{datasets}

