% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex_eqtl.R
\docType{data}
\name{gtex_eqtl}
\alias{gtex_eqtl}
\title{Data frame of eQTLs from GTEx study}
\format{A data frame with 187263 rows and 11 columns
\describe{
	\item{gene_name}{Gene name}
	\item{tissue}{Tissue}
	\item{gene_start}{Gene chromosome : gene start position (b37)}
	\item{SNP}{SNP rs ID}
	\item{snp_position}{SNP chromosome : SNP position (b37)}
	\item{effect_allele}{Effect allele}
	\item{other_allele}{Non effect allele}
	\item{beta}{Effect size}
	\item{se}{Standard error}
	\item{P_value}{p-value}
	\item{n}{samplesize}
}}
\source{
\url{https://scmv-ieugit.epi.bris.ac.uk/gh13047/gtex_data}
}
\usage{
gtex_eqtl
}
\description{
A total of 273700 Gene expression - SNP associations from 27094 gene names and 44 tissues. Obtained from \url{http://www.gtexportal.org/home/}.
These data only include associations that GTEx labelled as \code{is_choson_snp}.
}
\keyword{datasets}

