% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metab_qtls.R
\docType{data}
\name{metab_qtls}
\alias{metab_qtls}
\title{Data frame of QTLs of metabolites}
\format{A data frame with 1088 rows and 12 columns
     \describe{
             \item{phenotype}{phenotype}
             \item{chromosome}{chromosome}
             \item{position}{position}
             \item{SNP}{SNP rs ID}
             \item{effect_allele}{Effect allele}
             \item{other_allele}{Non effect allele}
             \item{eaf}{effect allele frequency}
             \item{beta}{Effect size}
             \item{se}{Standard error}
             \item{pval}{p-value}
             \item{n_studies}{number of studies}
             \item{n}{samplesize}
     }}
\source{
unpublished
}
\usage{
metab_qtls
}
\description{
Data provided by Peter Wurtz and Mika Ala-Korpela ahead of publication
}
\keyword{datasets}

