% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbase_instruments.R
\docType{data}
\name{mrbase_instruments}
\alias{mrbase_instruments}
\title{Data frame of GWAS significant SNPs obtained from the EBI GWAS catalog}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 41203 rows and 16 columns.}
\source{
\url{http://www.mrbase.org}
}
\usage{
mrbase_instruments
}
\description{
Created from the MR-Base database 22/12/2017
The top hits (5e-8) are extracted from these files and then the 1kg European dataset is used to clump the top hits (rsq = 0.01, kb = 10000)
}
\keyword{datasets}
