% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomic_qtls.R
\docType{data}
\name{proteomic_qtls}
\alias{proteomic_qtls}
\title{Data frame of proteomic QTLs from Deming et al 2016}
\format{
A data frame with 56 rows and 13 columns
\describe{
\item{analyte}{analyte}
\item{chr}{chr}
\item{position}{position}
\item{SNP}{SNP}
\item{gene}{gene}
\item{location}{location}
\item{annotation}{annotation}
\item{effect_allele}{effect_allele}
\item{other_allele}{other_allele}
\item{eaf}{eaf}
\item{maf}{MAF}
\item{pval}{pval}
\item{beta}{beta}
\item{se}{se}
}
}
\source{
\url{https://scmv-ieugit.epi.bris.ac.uk/gh13047/proteomic_qtls}
}
\usage{
proteomic_qtls
}
\description{
GWAS of immunoassays of 190 protein levels from plasma
}
\keyword{datasets}
