% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbase_instruments.R
\docType{data}
\name{mrbase_instruments}
\alias{mrbase_instruments}
\title{Data frame of GWAS significant SNPs obtained from the EBI GWAS catalog}
\format{A data frame with 41203 rows and 16 columns
\describe{
      \item{exposure}{This phenotype also has units and details if they are necessary to distinguish different studies}
      \item{id.exposure}{id.exposure}
      \item{samplesize.exposure}{samplesize.exposure}
      \item{ncase.exposure}{ncase.exposure}
      \item{ncontrol.exposure}{ncontrol.exposure}
      \item{mr_keep.exposure}{mr_keep.exposure}
      \item{pval_origin.exposure}{pval_origin.exposure}
      \item{data_source.exposure}{data_source.exposure}
      \item{SNP.exposure}{SNP.exposure}
      \item{effect_allele.exposure}{effect_allele.exposure}
      \item{other_allele.exposure}{other_allele.exposure}
      \item{beta.exposure}{beta.exposure}
      \item{se.exposure}{se.exposure}
      \item{pval.exposure}{pval.exposure}
      \item{units.exposure}{units.exposure}
      \item{eaf.exposure}{eaf.exposure}
}}
\source{
\url{http://www.mrbase.org}
}
\usage{
mrbase_instruments
}
\description{
Created from the MR-Base database 22/12/2017
The top hits (5e-8) are extracted from these files and then the 1kg European dataset is used to clump the top hits (rsq = 0.01, kb = 10000)
}
\keyword{datasets}
